+++ settings.js
@@ -17,7 +17,7 @@ export const magmaCoreProfitTrackerOverlayGui = new Gui();
 export const abandonedQuarryTrackerOverlayGui = new Gui();
 export const fishingProfitTrackerOverlayGui = new Gui();
 
-const categories = ["General", "Chat", "Alerts", "Overlays", "Items and storages", "Commands"]
+const categories = ["General", "Chat", "Alerts", "Overlays", "Items and storages", "World", "Commands"]
 const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
 
 .addButton({
@@ -1464,6 +1464,66 @@ ${RED}Hidden if you have no fishing rod in your hotbar!`,
     title: "Price per T1 attribute shard",
     description: "Render price per T1 attribute level in the auctioned Attribute Shard's lore, based on item's price. Helps to compare prices for high-tier attribute shards on AH.",
     subcategory: "Item lore"
+})
+
+.addTextParagraph({
+    category: "World",
+    configName: "worldBoxingText",
+    title: "Boxing",
+    description: `This section allows to draw boxes around some entities. ${BOLD}Boxes are not visible through walls!\n${RED}Hidden if you have no fishing rod in your hotbar!`,
+    centered: false,
+    subcategory: "Boxing"
+})
+.addSlider({
+    category: "World",
+    configName: "boxLineWidth",
+    title: "Box line width",
+    description: "The line width for boxing the entities.",
+    options: [1, 10],
+    value: 2,
+    subcategory: "Boxing"
+})
+.addSwitch({
+    category: "World",
+    configName: "boxWikiTiki",
+    title: "Box Wiki Tikis",
+    description: "Render box around Wiki Tikis nearby.",
+    subcategory: "Boxing"
+})
+.addSwitch({
+    category: "World",
+    configName: "boxWikiTikiLaserTotem",
+    title: "Box Wiki Tiki Laser Totems",
+    description: "Render box around Wiki Tiki Laser Totems nearby.",
+    subcategory: "Boxing"
+})
+.addSwitch({
+    category: "World",
+    configName: "boxBlueRingedOctopus",
+    title: "Box Blue Ringed Octopuses",
+    description: "Render box around Blue Ringed Octopuses nearby.",
+    subcategory: "Boxing"
+})
+.addSwitch({
+    category: "World",
+    configName: "boxTitanoboaHead",
+    title: "Box Titanoboa Heads",
+    description: "Render box around Titanoboa Heads nearby.",
+    subcategory: "Boxing"
+})
+.addSwitch({
+    category: "World",
+    configName: "boxFieryScuttler",
+    title: "Box Fiery Scuttlers",
+    description: "Render box around Fiery Scuttlers nearby.",
+    subcategory: "Boxing"
+})
+.addSwitch({
+    category: "World",
+    configName: "boxJawbusFollowers",
+    title: "Box Jawbus Followers",
+    description: "Render box around Jawbus Followers nearby.",
+    subcategory: "Boxing"
 })
 
 .addButton({