+++ docs/CHANGELOG.md
@@ -5,6 +5,10 @@
 Released: ???
 
 Features:
+- Added option to box some entities hardly visible in the world [disabled by default]:
+  - Boxing added for Wiki Tiki, Wiki Tiki Laser Totem, Blue Ringed Octopus, Titanoboa's Head, Fiery Scuttler, Jawbus Follower.
+  - Rendered box is not visible through walls.
+  - Observe new settings category named World, to enable.
 
 Bugfixes: